wms/sys11: clean up current_ball/match n/a values
Match never uses a BCD "F" in the lower digit, so we don't need to check
for it.  Also, the Python processor performs the BCD conversion before
checking for special values, and treats a non-digit (A-F) as 0.

When looking for a "ball" displayed instead of "match", we just need to
check for a non-zero lower digit (ball 1-5).